Moved classes responsible for migation to one dir
