+++ collect_app/src/main/java/org/odk/collect/android/activities/MainMenuActivity.java
@@ -57,7 +57,7 @@ import org.odk.collect.android.preferences.PreferencesActivity;
 import org.odk.collect.android.preferences.Transport;
 import org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns;
 import org.odk.collect.android.storage.StorageInitializer;
-import org.odk.collect.android.storage.StorageMigrationViewModel;
+import org.odk.collect.android.storage.migration.StorageMigrationViewModel;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageStateProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
